requirements.txt
sentence-transformers
faiss-cpu
pandas
torch
os
-q google.generativeai --upgrade
Flask
pandas
json
